+++ src/background/messaging/messaging.ts
@@ -365,8 +365,9 @@ const messagesHandler = async (message: Message, sender: Runtime.MessageSender) 
 
             return vpnProvider.requestSupport(reportData);
         }
-        case MessageType.OPEN_PREMIUM_PROMO_PAGE: {
-            return actions.openPremiumPromoPage();
+        case MessageType.OPEN_FORWARDER_URL_WITH_EMAIL: {
+            const { forwarderUrlQueryKey } = data;
+            return actions.openForwarderUrlWithEmail(forwarderUrlQueryKey);
         }
         case MessageType.SET_FLAG: {
             const { key, value } = data;