+++ src/background/networkConnectionObserver/networkConnectionObserverMv2.ts
@@ -2,7 +2,16 @@
  * Class observes network state
  */
 export class NetworkConnectionObserver {
+    private readonly onlineHandler: () => Promise<void>;
+
     constructor(callback: () => Promise<void>) {
-        window.addEventListener('online', callback);
+        this.onlineHandler = callback;
+    }
+
+    /**
+     * Initializes the network connection observer.
+     */
+    public init() {
+        window.addEventListener('online', this.onlineHandler);
     }
 }