+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- TODO: Add diff link -->
+## Unreleased
+
+### Added
+
+- Limit on the number of characters a user can enter when adding a custom DNS server [#169].
+
+### Changed
+
+- Improved validation of domain name when adding an exclusions [#191].
+
+### Fixed
+
+- Inconsistent colors of UI elements [#171].
+- Disable resend email button for 60 seconds after clicking it [#192].
+
+[#169]: https://github.com/AdguardTeam/AdGuardVPNExtension/issues/169
+[#191]: https://github.com/AdguardTeam/AdGuardVPNExtension/issues/191
+[#171]: https://github.com/AdguardTeam/AdGuardVPNExtension/issues/171
+[#192]: https://github.com/AdguardTeam/AdGuardVPNExtension/issues/192
+
 <!-- TODO: Add release date -->
 ## [2.4.1]